+++ src/main/java/ui/category/ColourPalette.java
@@ -1,4 +1,4 @@
-package ui;
+package ui.category;
 
 import java.awt.*;
